deleting old DraggableArea component
