+++ theatre/studio/src/panels/SequenceEditorPanel/RightOverlay/HorizontalScrollbar.tsx
@@ -1,5 +1,5 @@
 import type {SequenceEditorPanelLayout} from '@theatre/studio/panels/SequenceEditorPanel/layout/layout'
-import DraggableArea from '@theatre/studio/uiComponents/DraggableArea2'
+import DraggableArea from '@theatre/studio/uiComponents/DraggableArea'
 import {useVal} from '@theatre/react'
 import type {IRange} from '@theatre/shared/utils/types'
 import type {Pointer} from '@theatre/dataverse'
