+++ theatre/studio/src/uiComponents/form/BasicNumberInput.tsx
@@ -2,7 +2,7 @@ import React, {useMemo, useRef, useEffect} from 'react'
 import type {MutableRefObject} from 'react'
 import {clamp, isInteger, round} from 'lodash-es'
 import styled from 'styled-components'
-import DraggableArea from '@theatre/studio/uiComponents/DraggableArea2'
+import DraggableArea from '@theatre/studio/uiComponents/DraggableArea'
 import mergeRefs from 'react-merge-refs'
 import useRefAndState from '@theatre/studio/utils/useRefAndState'
 import useOnClickOutside from '@theatre/studio/uiComponents/useOnClickOutside'
